Add field min_run_time to energy systems that support strategy storage_driven.

+++ src/energy_systems/heat_producers/gas_boiler.jl
@@ -15,7 +15,7 @@ mutable struct GasBoiler <: ControlledSystem
 
     power :: Float64
     min_power_fraction :: Float64
-    min_run_time :: Float64
+    min_run_time :: UInt
 
     function GasBoiler(uac :: String, config :: Dict{String, Any})
         return new(
